% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply2TrajId}
\alias{apply2TrajId}
\title{Apply function to each pair of trajectories from two Trajs objects.}
\usage{
apply2TrajId(trajs1, trajs2, fun, ..., simplify = FALSE)
}
\arguments{
\item{trajs1, trajs2}{Trajs objects.}

\item{fun}{A function that maps two Trajs objects with a single identical
trajsId to an arbitrary value.}

\item{...}{Further arguments for \code{fun}.}

\item{simplify}{Should the resulting list of values be simplified to an array
if possible?}
}
\value{
A list or an array depending on \code{simplify}. The trajIds are the names
of the list entries or the names of the last dimension of the array.
}
\description{
Applies the function \code{fun} to each pair of trajectories with the same
\code{trajId} of two Trajs objects. Only trajIds that occur in both objects
are used.
}
