% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{applyToEachDimAndId}
\alias{applyToEachDimAndId}
\title{Apply function columns of Trajs state and obtain a new Trajs.}
\usage{
applyToEachDimAndId(trajs, func, outTime, ...)
}
\arguments{
\item{trajs}{A Trajs object.}

\item{func}{A function that takes the time vector of \code{trajs}, a column vector
of the state of a single trajectory, the expected times for the resulting
Trajs, and possibly further arguments, It returns a Trajs (with arbitrary
trajId as this will be overwritten).}

\item{outTime}{A numeric vector of times for the output Trajs.}

\item{...}{further arguments to \code{func}}
}
\value{
A Trajs object.
}
\description{
For each column and trajId, apply a function and return a Trajs with
specified times.
}
