% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{applyToTrajStateCols}
\alias{applyToTrajStateCols}
\title{Apply function columns of trajs state}
\usage{
applyToTrajStateCols(trajs, func, ...)
}
\arguments{
\item{trajs}{A Trajs object.}

\item{func}{A function that takes a column vector of the state of a single
trajectory and returns a vector or a tibble.}

\item{...}{further arguments to \code{func}}
}
\value{
A tibble with a column trajId and further columns containing the
results of \code{func}
}
\description{
For each column and trajId, apply a function and return a tibble of the
results.
}
