% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{applyTrajId}
\alias{applyTrajId}
\title{Apply function to each trajectory of a Trajs object.}
\usage{
applyTrajId(trajs, fun, ..., simplify = FALSE)
}
\arguments{
\item{trajs}{A Trajs object.}

\item{fun}{A function that maps a Trajs object with a single trajsId to
an arbitrary value.}

\item{...}{Further arguments for \code{fun}.}

\item{simplify}{Should the resulting list of values be simplified to an array
if possible?}
}
\value{
A list or an array depending on \code{simplify}. The trajIds are the names
of the list entries or the names of the last dimension of the array.
}
\description{
Applies the function \code{fun} to each trajectory identified by the column trajId
of the Trajs object \code{trajs}.
}
