% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{calculateProjection}
\alias{calculateProjection}
\title{Calculate linear projection from data}
\usage{
calculateProjection(y, dim)
}
\arguments{
\item{y}{A numeric matrix.}

\item{dim}{A single positive integer. The target dimension.}
}
\value{
A named list of two function: \code{project(x)} maps \code{d} dimensional data
stored in a matrix \code{x} with \code{d} columns to the lower dimensional space and
outputs the respective values as matrix with the same number of rows as the
input and \code{dim} columns. \code{embed(x)} maps \code{dim} dimensional data back to the
\code{d} dimensional space, again via matrices with the respective column count.
}
\description{
Calculates a PCA based linear projection to a \code{dim} space from a
\code{d = ncol(y)} dimensional space from \code{n = nrow(y)} data point stored in \code{y}.
}
