% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{compareTrajStates}
\alias{compareTrajStates}
\title{Calls a function and the two state matrices of the trajs and returns the resulting trajs of the same times but potentially of different dimension depending on the function.
If timeRange and nBasePoints are give an interpoation is done before the function is applied.}
\usage{
compareTrajStates(
  traj1,
  traj2,
  stateCompFun,
  timeRange = NULL,
  nBasePoints = 1000
)
}
\description{
Calls a function and the two state matrices of the trajs and returns the resulting trajs of the same times but potentially of different dimension depending on the function.
If timeRange and nBasePoints are give an interpoation is done before the function is applied.
}
