% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEBtrajs-package.R
\docType{package}
\name{DEEBtrajs-package}
\alias{DEEBtrajs}
\alias{DEEBtrajs-package}
\title{DEEBtrajs: S3 Class for Multiple Trajectories}
\description{
A S3 class Trajs is defined. It stores data from multiple trajectories (depending on time) of a dynamical system that depends.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
