% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{fillNaState}
\alias{fillNaState}
\title{Fills all state rows with at least one NA entry with a given (single) state.}
\usage{
fillNaState(trajs, state)
}
\description{
Fills all state rows with at least one NA entry with a given (single) state.
}
