% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getClosestInTime}
\alias{getClosestInTime}
\title{Get Trajs with States that are Closest To Given Time Point}
\usage{
getClosestInTime(trajs, time)
}
\arguments{
\item{trajs}{A Trajs object.}

\item{time}{A single finite number.}
}
\value{
A Trajs object with one row for each trajId of the input.
}
\description{
Get Trajs with States that are Closest To Given Time Point
}
