% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{getIdentityProjection}
\alias{getIdentityProjection}
\title{Return the identity projection.}
\usage{
getIdentityProjection()
}
\value{
A named list of two function: \code{project(x)} and \code{embed(x)}, both
realize the identity function.
}
\description{
Return the identity projection.
}
