% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getInitialState}
\alias{getInitialState}
\title{Get Trajs with (Potentially Interpolated) States at Initial Time}
\usage{
getInitialState(trajs, initialTime = 0)
}
\arguments{
\item{trajs}{A Trajs object.}

\item{initialTime}{A single finite number.}
}
\value{
A Trajs object with one row for each trajId of the input.
}
\description{
Get Trajs with (Potentially Interpolated) States at Initial Time
}
