% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeStep.R
\name{getTimeStepTrajs}
\alias{getTimeStepTrajs}
\title{Get the (average) time step length of a DEEBtrajs object.}
\usage{
getTimeStepTrajs(trajs, requireConst = TRUE)
}
\arguments{
\item{trajs}{A DEEBtrajs object with one or many trajectories.}

\item{requireConst}{If \code{TRUE}, an error is raised if time steps differ between
steps or trajectories.}
}
\description{
Returns the mean time step over all trajectories and steps.
}
