% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTrajIds}
\alias{getTrajIds}
\title{Return all unique trajIds}
\usage{
getTrajIds(trajs)
}
\arguments{
\item{trajs}{A Trajs object.}
}
\value{
An integer vector of trajIds.
}
\description{
Return all unique trajIds
}
