% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeDerivTrajs}
\alias{makeDerivTrajs}
\title{Create a DerivTrajs Object}
\usage{
makeDerivTrajs(state, deriv = NULL)
}
\arguments{
\item{state}{A numeric matrix with 1 or n rows.}

\item{deriv}{\code{NULL} or a numeric matrix with 1 or n rows.}
}
\value{
An object of S3 class DerivTrajs.
}
\description{
Create a DerivTrajs Object
}
