% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeTrajs}
\alias{makeTrajs}
\title{Create a Trajs Object}
\usage{
makeTrajs(time = NULL, state, trajId = NULL, deriv = NULL)
}
\arguments{
\item{time}{\code{NULL} or a numeric vector of length 1 or n.}

\item{state}{A numeric matrix with 1 or n rows.}

\item{trajId}{\code{NULL} or an integer vector with 1 or n rows.}

\item{deriv}{\code{NULL} or a numeric matrix with 1 or n rows.}
}
\value{
An object of S3 class Trajs.
}
\description{
Create a Trajs Object
}
