% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{makeTrajsStateConst}
\alias{makeTrajsStateConst}
\title{Applies function to traj state cols and returns a trajs with the same time and dimension but a constant state.}
\usage{
makeTrajsStateConst(traj, fun)
}
\description{
Applies function to traj state cols and returns a trajs with the same time and dimension but a constant state.
}
