% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{mapTrajs2Trajs}
\alias{mapTrajs2Trajs}
\title{Apply function to each trajectory of a Trajs object to obtain a new Trajs.}
\usage{
mapTrajs2Trajs(trajs, fun, ...)
}
\arguments{
\item{trajs}{A Trajs object.}

\item{fun}{A function that maps a Trajs object with a single trajsId to an
arbitrary value.}

\item{...}{Further arguments for \code{fun}.}
}
\value{
A Trajs object with the same trajIds as \code{trajs} but potentially
different times.
}
\description{
Applies the function \code{fun} to each trajectory identified by the column trajId
of the DEEBtrajs object \code{trajs}. The result of \code{fun} is converted to a Trajs
and the respective trajId is set. The collected results are returned as a
single Trajs.
}
