% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{applyToTrajStateCols}
\alias{applyToTrajStateCols}
\title{trajs to tibble with columns trajId and value}
\usage{
applyToTrajStateCols(trajs, func)
}
\description{
trajs to tibble with columns trajId and value
}
