% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeStep.R
\name{getTimeStepTrajs}
\alias{getTimeStepTrajs}
\title{Get the time step length of a DEEBtrajs object.}
\usage{
getTimeStepTrajs(trajs)
}
\arguments{
\item{trajs}{A DEEBtrajs object with one or many trajectories.}
}
\description{
Assuming there is a single time step length for all trajectories in the a
trajs object, this time step is returned. Otherwise an error is thrown.
}
