% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeDerivTrajs}
\alias{makeDerivTrajs}
\title{Create a DerivTrajs Object}
\usage{
makeDerivTrajs(state, deriv)
}
\arguments{
\item{state}{A numeric matrix with 1 or n rows.}

\item{deriv}{A numeric matrix with 1 or n rows.}
}
\value{
An object of S3 class Trajs.
}
\description{
Create a DerivTrajs Object
}
